refactor: useCart 코드 정리
- 일부 코드 모듈화
- 과도한 연산을 요구하는 함수의 역할을 이전

등
