+++ src/components/CartList/CartList.tsx
@@ -1,20 +1,19 @@
-import {useRecoilValue} from "recoil";
-import useCart from "../../hooks/useCart";
-import {cartCountSelector} from "../../recoil/cartAtoms";
+import { useRecoilValue } from "recoil";
+import { cartCountSelector, cartState } from "../../recoil/cartAtoms";
 import CartItem from "../CartItem";
-import {CartListCheckCounter, CartListController, CartListTitle} from "./CartList.style";
+import { CartListCheckCounter, CartListController } from "./CartList.style";
 
 function CartList() {
-  const {cartList} = useCart();
+  const cartList = useRecoilValue(cartState);
   const cartCount = useRecoilValue(cartCountSelector);
 
   return (
-    <div style={{width: '100%'}}>
+    <div style={{ width: '100%' }}>
       {cartList.map((cart) => (
-        <CartItem key={cart.id} cart={cart}/>
+        <CartItem key={cart.id} cart={cart} />
       ))}
       <CartListController>
-        <input type='checkbox'/>
+        <input type='checkbox' />
         <CartListCheckCounter>전체선택 (2/{cartCount})</CartListCheckCounter>
         <div>
           <button>선택삭제</button>
