+++ src/hooks/useCart.ts
@@ -1,20 +1,19 @@
-import {useRecoilState} from 'recoil';
-import type {CartItem, NewCartItem, ProductItem} from '../types/types';
-import {cartState} from '../recoil/cartAtoms';
-import {fetchAddCart, fetchDeleteCart, fetchUpdateCart} from '../api/api';
+import { useRecoilState } from 'recoil';
+import type { NewCartItem, ProductItem } from '../types/types';
+import { cartState } from '../recoil/cartAtoms';
+import { fetchAddCart, fetchDeleteCart, fetchUpdateCart } from '../api/api';
+import { updateCartListQuantity } from '../domain/cart';
 
 function useCart() {
   const [cartList, setCartList] = useRecoilState(cartState);
 
-  const getQuantityByProductId = (id: number) => {
-    const targetCart = cartList.find((cart) => cart.id === id);
-    return targetCart?.quantity ?? 0;
-  };
-
   const addCart = (product: ProductItem) => {
-    if (cartList.find((cartItem) => cartItem.id === product.id) === undefined) {
-      const newCartItem: NewCartItem = {id: product.id, quantity: 1, product};
-      setCartList([...cartList, newCartItem]);
+    const isCartItemExist = cartList.some((cartItem) => cartItem.id === product.id);
+
+    if (!isCartItemExist) {
+      const newCartItem: NewCartItem = { id: product.id, quantity: 1, product };
+      const updatedCartList = [...cartList, newCartItem];
+      setCartList(updatedCartList);
       fetchAddCart(newCartItem.id);
     }
   };
@@ -23,17 +22,6 @@ function useCart() {
     const removedCartList = cartList.filter((cart) => cart.id !== id);
     setCartList(removedCartList);
     fetchDeleteCart(id);
-  };
-
-  const updateCartListQuantity = (cartList: CartItem[], id: number, newQuantity: number) => {
-    const targetIndex = cartList.findIndex(cartItem => cartItem.id === id);
-    const targetCart = cartList[targetIndex];
-    const updatedCart = {
-      ...targetCart,
-      quantity: newQuantity
-    };
-    cartList[targetIndex] = updatedCart;
-    return cartList;
   };
 
   const setCartQuantity = (id: number, quantity: number) => {
@@ -46,20 +34,9 @@ function useCart() {
     }
   };
 
-  const increaseCart = (id: number) => {
-    setCartQuantity(id, getQuantityByProductId(id) + 1);
-  };
-
-  const decreaseCart = (id: number) => {
-    setCartQuantity(id, getQuantityByProductId(id) - 1);
-  };
-
   return {
-    cartList,
-    getQuantityByProductId,
     addCart,
-    increaseCart,
-    decreaseCart,
+    removeCart,
     setCartQuantity,
   };
 }
